refactor(node): extract proof_history ExEx into exex.rs
